requirements.txt
google-cloud-storage
google-cloud-bigquery
python-dotenv
scikit-learn
pandas
db-dtypes
joblib
matplotlib
seaborn
